fiducial: marker map owns the tag size
